+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ### Changelog
 
 #### Version - TBD
+* Updated GameFinder to 4.1.0
+  * Fixes Wabbajack not Launching on some systems
+* Added the GameFinder module for EA Desktop
+  * Only `Dragon Age: Origins` & `Dragon Age: Inquisition` supported for now.
+  * For other games we still need to collect the right store IDs
 * Fixed Final Fantasy 7: Remake Intergrade meta data
 
 #### Version - 3.4.1.0 - 12/21/2023